Create README.md
Created a Proccessing solution to find Tivaware Launchpad examples on the LAN. There is a function in Locator.c the listens for broadcasts and returns a unicast with the Launchpad's IP address (and other information) which this works with. 

Tested on Tivaware version TivaWare_C_Series-2.1.4.178 locator. 

See traffic regarding my request for a TI client here: https://e2e.ti.com/support/microcontrollers/tiva_arm/f/908/p/690625/2562207#2562207

Thanks to Nathan Martin for helping me with the udp.setReceiveHandler("myCustomReceiveHandler"); and writing the handler.